+++ src/Page/Exception/PageEditException.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Message\Mothership\CMS\Page\Exception;
+
+/**
+ * Class PageEditException
+ *
+ * @author Thomas Marchant <thomas@mothership.ec?
+ */
+class PageEditException extends \LogicException
+{
+
+}